Fix sms json parse error (#88)
When sending an SMS message in the SMS plugin, the message is sent but
the CLI returns the following error:

`SyntaxError: Unexpected token o in JSON at position 1`

The plugin was trying to `JSON.parse()` the response from `vetch`, which
automatically parses responses. This just removes the additional
parsing.

This also has some version bumps that were missing that I apparently had
locally.