build: fixing typescript (#90)
Typescript failed subsequent builds due to the `build` script for
`@vonage/cli-utils` removing `dist` but not `tsconfig.buildinfo`. Since
utils is a shared package, the `composite` flag must be true. Having
that tells typescript to cache the build if no files have changed. Since
the build was removing the dist folder, TS was not re-building. There is
no need to clean the dist folder before building the packages